// drebin_store/ClientApp/src/app/models/last-seen-codec-entry.ts
import { MainQuestStage } from './main-quest-stage';

export class LastSeenCodecEntry {
    mainQuestStage: MainQuestStage;
    numberOfQuest: number;
    numberOfCall: number;

    // constructor(mainQuestStage: MainQuestStage, numberOfQuest: number, numberOfCall: number) {
    //     this.mainQuestStage = mainQuestStage;
    //     this.numberOfQuest = numberOfQuest;
    //     this.numberOfCall = numberOfCall;
    // }
}